requirements.txt
discord.py
selenium
flask
pytz
gtts
webdriver-manager